move integrations to separate subdir + make test_loads more versatile
