+++ pwndbg/integration/binja.py
