+++ tests/gdb-tests/tests/utils.py
@@ -57,4 +57,8 @@ def run_gdb_with_script(
         output,
     )
 
+    # It also shows every single registered function, so we change it to xxx
+    # so as to not break this test every time a new function is added
+    output = re.sub(r"created (\$\w+, )*\$\w+ GDB functions", r"created xxx GDB functions", output)
+
     return output
